cast lum to int

+++ gateComs/LEDUtils.py
@@ -217,5 +217,5 @@ class LEDStrip:
         x = self.updateFrame(255)
         for j in range(self.strip.numPixels()):
             lum = ((math.sin((x*math.pi*0.02)-(math.pi*.5))+1)*127.5)
-            self.strip.setPixelColor(j, Color(lum,lum,lum))
+            self.strip.setPixelColor(j, Color((int)lum,(int)lum,(int)lum))
         self.strip.show()
